Moved search-settings
