+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/search-settings/index.tsx
