+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/search-settings/search-settings.tsx
@@ -15,22 +15,21 @@
 import * as React from 'react'
 import withListenTo, {
   WithBackboneProps,
-} from '../../container/backbone-container'
-const user = require('../../../component/singletons/user-instance.js')
-const properties = require('../../../js/properties.js')
-const Property = require('../../../component/property/property.js')
-const PropertyView = require('../../../component/property/property.view.js')
-import MarionetteRegionContainer from '../../../react-component/container/marionette-region-container'
-const QuerySettingsView = require('../../../component/query-settings/query-settings.view.js')
-const QueryModel = require('../../../js/model/Query.js')
-const ConfirmationView = require('../../../component/confirmation/confirmation.view.js')
+} from '../container/backbone-container'
+const user = require('../../component/singletons/user-instance.js')
+const properties = require('../../js/properties.js')
+const Property = require('../../component/property/property.js')
+const PropertyView = require('../../component/property/property.view.js')
+import MarionetteRegionContainer from '../../react-component/container/marionette-region-container'
+const QuerySettingsView = require('../../component/query-settings/query-settings.view.js')
+const QueryModel = require('../../js/model/Query.js')
+const ConfirmationView = require('../../component/confirmation/confirmation.view.js')
 import styled from 'styled-components'
 import { hot } from 'react-hot-loader'
 
 const Root = styled.div`
   overflow: hidden;
-  padding: ${props => props.theme.minimumSpacing}
-    ${props => props.theme.minimumSpacing};
+  padding: ${props => props.theme.minimumSpacing};
 `
 
 const PropertyResultCount = styled.div`
